first stab at rewriting recipe-run page
